+++ ask-smapi-model/ask_smapi_model/v1/skill/interaction_model/jobs/__init__.py
@@ -1,0 +1,39 @@
+# coding: utf-8
+
+#
+# Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the 'License'). You may not use this file
+# except in compliance with the License. A copy of the License is located at
+#
+# http://aws.amazon.com/apache2.0/
+#
+# or in the 'license' file accompanying this file. This file is distributed on an 'AS IS' BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for
+# the specific language governing permissions and limitations under the License.
+#
+from __future__ import absolute_import
+
+from .execution import Execution
+from .trigger import Trigger
+from .catalog_auto_refresh import CatalogAutoRefresh
+from .list_job_definitions_response import ListJobDefinitionsResponse
+from .reference_version_update import ReferenceVersionUpdate
+from .job_api_pagination_context import JobAPIPaginationContext
+from .job_error_details import JobErrorDetails
+from .update_job_status_request import UpdateJobStatusRequest
+from .create_job_definition_response import CreateJobDefinitionResponse
+from .job_definition_status import JobDefinitionStatus
+from .validation_errors import ValidationErrors
+from .get_executions_response import GetExecutionsResponse
+from .catalog import Catalog
+from .resource_object import ResourceObject
+from .job_definition import JobDefinition
+from .job_definition_metadata import JobDefinitionMetadata
+from .interaction_model import InteractionModel
+from .scheduled import Scheduled
+from .create_job_definition_request import CreateJobDefinitionRequest
+from .referenced_resource_jobs_complete import ReferencedResourceJobsComplete
+from .slot_type_reference import SlotTypeReference
+from .dynamic_update_error import DynamicUpdateError
+from .execution_metadata import ExecutionMetadata